Refactor knot vector and basis functions.

+++ src/baf/zero_degree_b_spline_basis_function.cc
@@ -19,8 +19,8 @@ You should have received a copy of the GNU Lesser General Public License along w
 baf::ZeroDegBSplBasFnc::ZeroDegBSplBasFnc(const baf::KnotVector &knot_vector,
                                           const KnotSpan &start_of_support) :
     BasisFunction(knot_vector, Degree{0}, start_of_support),
-    start_knot_(knot_vector.GetKnot(start_of_support.get())),
-    end_knot_(knot_vector.GetKnot(start_of_support.get() + 1)) {}
+    start_knot_(knot_vector.GetKnot(static_cast<size_t>(start_of_support.get()))),
+    end_knot_(knot_vector.GetKnot(static_cast<size_t>(start_of_support.get()) + 1)) {}
 
 double baf::ZeroDegBSplBasFnc::EvaluateOnSupport(const ParamCoord /* param_coord*/&) const {
   return util::NumericSettings<double>::AreEqual(start_knot_.get(), end_knot_.get()) ? 0.0 : 1.0;
